requirements.txt
highrise-bot-sdk==24.1.0
flask==3.0.3
supabase==2.28.3
python-dotenv==1.2.2
pytz==2026.1.post1
certifi==2026.2.25
requests==2.33.0